backend/requirements.txt
Flask
flask-cors
supabase
zappa
setuptools
flask-supabase
python-dotenv
pytest-cov
